Created UserProfiling component
