+++ src/components/CountryCheckbox.vue
@@ -1,7 +1,8 @@
 <template>
-  <div v-show="country.toLowerCase().includes(search.toLowerCase())" class="flex has-[:checked]:bg-blue-100">
-    <input @change="$emit('toggle-country', $event.target.value)" :checked="selectedCountries.includes(country)" type="checkbox" :id="'checkbox' + country" :value="country" class="self-start mt-[5px]" />
-    <label :for="'checkbox' + country" class="ml-2 text-left w-full select-none">{{ country }}</label>
+  <div v-show="country.toLowerCase().includes(search.toLowerCase())" class="flex has-[:checked]:bg-blue-100 cursor-pointer">
+    <input v-if="selectedCountries.length < 11" @change="$emit('toggle-country', $event.target.value)" :checked="selectedCountries.includes(country)" type="checkbox" :id="'checkbox' + country" :value="country" class="self-start mt-[5px] cursor-pointer" />
+    <input v-else @click.prevent @click="$emit('toggle-country', $event.target.value)" :checked="selectedCountries.includes(country)" type="checkbox" :id="'checkbox' + country" :value="country" class="self-start mt-[5px] cursor-pointer" />
+    <label :for="'checkbox' + country" class="ml-2 text-left w-full select-none cursor-pointer">{{ country }}</label>
   </div>
 </template>
